+++ src/app/providers/ThemeProvider/index.ts
@@ -1,0 +1,4 @@
+import { useTheme } from './lib/useTheme'
+import { ThemeProvider } from './ui/ThemeProvider'
+
+export { ThemeProvider, useTheme }
