+++ src/app/styles/index.scss
@@ -1,7 +1,7 @@
 @import "reset";
-@import "variables/global";
-@import "themes/normal";
-@import "themes/dark";
+@import "./variables/global";
+@import "./themes/normal";
+@import "./themes/dark";
 
 .app {
   font: var(--font-m);
