--- src/components/Counter.module.scss
@@ -1,5 +1,0 @@
-.counter{
-  padding: 20px;
-  color: red;
-  background: #0f4;
-}
