+++ src/index.tsx
@@ -1,8 +1,8 @@
 import {render} from 'react-dom'
-import App from './App'
-import './styles/index.scss'
-import {BrowserRouter} from 'react-router-dom';
-import {ThemeProvider} from './theme/ThemeProvider';
+import App from './app/App'
+import './app/styles/index.scss'
+import {BrowserRouter} from 'react-router-dom'
+import {ThemeProvider} from 'app/providers/ThemeProvider'
 
 render(
   <BrowserRouter>
