+++ src/pages/AboutPage/ui/AboutPage.async.tsx
