+++ src/pages/AboutPage/ui/AboutPage.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 const AboutPage = () => {
   return (
     <div>
