+++ src/pages/MainPage/index.ts
@@ -1,0 +1,1 @@
+export {MainPageAsync as MainPage} from './ui/MainPage.async'
