+++ src/pages/MainPage/ui/MainPage.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 const MainPage = () => {
   return (
     <div>
