+++ src/shared/lib/classNames/classNames.ts
@@ -1,0 +1,15 @@
+export type Mods = Record<string, string | boolean | undefined>;
+
+export function classNames(
+  cls: string,
+  mods: Mods = {},
+  additional: Array<string | undefined> = [],
+): string {
+  return [
+    cls,
+    ...additional.filter(Boolean),
+    ...Object.entries(mods)
+      .filter(([_, value]) => Boolean(value))
+      .map(([className]) => className),
+  ].join(' ')
+}
