+++ tsconfig.json
@@ -9,7 +9,10 @@
     "moduleResolution": "node",
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
-//    "baseUrl": "."
+    "baseUrl": "./",
+    "paths": {
+      "*": ["./src/*"]
+    }
   },
   "ts-node": {
     "compilerOptions": {
