+++ webpack.config.ts
@@ -8,7 +8,8 @@ export default (env: BuildEnv) => {
   const paths: BuildPaths = {
     build: path.resolve(__dirname, 'build'),
     entry: path.resolve(__dirname, 'src', 'index.tsx'),
-    html: path.resolve(__dirname, 'public', 'index.html')
+    html: path.resolve(__dirname, 'public', 'index.html'),
+    src: path.resolve(__dirname, 'src')
   }
 
   const PORT = env.port || 3003
